+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod threadpool;
+mod threadpool;
 pub mod graph;
 pub use graph::{Graph, Recipe};
 pub use threadpool::ThreadExecute;